Issue_1146 Removed org/csstudio/archive/engine/model/VTypeHelper
